web-servers/service-discovery/requirements.txt
kazoo
python-dotenv
